refactor(common): MongoDBException pacakge
