+++ alarmeet-api/src/main/java/click/alarmeet/alarmeetapi/apis/groups/repository/GroupRepositoryCustom.java
@@ -4,7 +4,7 @@ import java.util.Map;
 
 import org.bson.types.ObjectId;
 
-import click.alarmeet.alarmeetcommon.mongodb.MongoDBException;
+import click.alarmeet.alarmeetcommon.exception.mongodb.MongoDBException;
 
 public interface GroupRepositoryCustom {
 	/**
